requirements.txt
jupyterlab
notebook
pandas
numpy
matplotlib
seaborn
scikit-learn
pdfplumber
plotly
python-dotenv
pymongo
streamlit as st
google.generativeai